pyproject.toml
[tool.poetry]
name = "python-reqres-in"
version = "0.1.0"
description = ""
authors = ["Vladislav <djefry.kamen@gmail.com>"]
readme = "README.md"
packages = [{include = "python_reqres_in"}]

[tool.poetry.dependencies]
python = "^3.10.5"
requests = "==2.28.2"
pytest = "==7.2.2"
voluptuous = "==0.13.1"
pytest-voluptuous = "==1.2.0"
selene = {version = "==2.0.0b17", allow-prereleases = true}
allure-pytest = "==2.13.1"
python-dotenv = "==1.0.0"
curlify = "==2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "-s --alluredir=allure-results --allure-no-capture"
